writing-practice/requirements.txt
streamlit
pillow
requests
openai
python-dotenv
manga-ocr
numpy==2.2.4